DND: Strip JIT lib on AIX prod builds
This roughly corresponds to what's done on Linux.

A better option might be to use the -bstabsplit linker option, which
will generate a .stab file with the separated debug info that DBX can
consume. (Note that DBGSUFF must be DBGSUFF=.stab and not DBGSUFF?=.stab
since the linker doesn't allow you to specify the suffix and DBX assumes
the suffix by default so it's best to set this in stone.)

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>